Safari-611.1.10/LayoutTests/editing/selection/selection-change-in-blur-event-by-remove-children-expected.txt

PASS selection.anchorOffset is 0
PASS selection.focusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
